+++ .github/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Add email blast marketing email
 -   Add 3 days left marketing email
+-   Add check to ensure application creation deadline has not passed yet
 
 ### Fixed